helloworld-dev-conference-langserve-ai/pyproject.toml
[tool.poetry]
name = "helloworld-dev-conference-langserve-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = ">2"
langchain-openai = "^0.1.23"
qdrant-client = "^1.11.1"
langchain-community = "^0.2.16"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
